+++ packages/guide/src/app/changelog/changelog-dialog/changelog-dialog.component.ts
@@ -1,0 +1,42 @@
+import { Component, inject } from '@angular/core';
+import {
+    MAT_DIALOG_DATA,
+    MatDialogActions,
+    MatDialogClose,
+    MatDialogContent,
+    MatDialogTitle
+} from "@angular/material/dialog";
+import { LatestChangelog } from "../latest-changelog.type";
+import { MarkdownComponent } from "ngx-markdown";
+import { MatCheckbox, MatCheckboxChange } from "@angular/material/checkbox";
+import { SettingsService } from "../../shared/services/settings.service";
+
+@Component({
+    selector: 'app-changelog-dialog',
+    standalone: true,
+    imports: [
+        MatDialogActions,
+        MatDialogClose,
+        MatDialogTitle,
+        MatDialogContent,
+        MarkdownComponent,
+        MatCheckbox
+    ],
+    templateUrl: './changelog-dialog.component.html',
+    styles: `
+        :host {
+            --mdc-dialog-subhead-weight: 700;
+            --mdc-dialog-subhead-size: 1.5rem;
+            --mdc-dialog-subhead-line-height: 1.8rem;
+
+        }`
+})
+export class ChangelogDialogComponent {
+    dialogData = inject<{ changelog: LatestChangelog }>(MAT_DIALOG_DATA);
+
+    #settingsService = inject(SettingsService)
+
+    toggleDisableChangelog($event: MatCheckboxChange) {
+        this.#settingsService.saveSettings({...this.#settingsService.getSettings(), disableChangelogs: $event.checked})
+    }
+}