+++ packages/guide/src/app/shared/interfaces/settings.interface.ts
@@ -4,4 +4,5 @@ export interface Settings {
     version: number;
     useBeta: boolean;
     language: AvailableLanguage;
+    disableChangelogs?: boolean
 }